exec: adds an option to log the command line
Having a way to log the full command line that is actually being
executed is of huge help in debugging.  When we know what command
line failed, we have a chance to try and reproduce the error ourselves.

Calls to os/exec.Command{,Context} are moved to their own package to
provide a drop-in logging wrapper, but also to allow independent setting
of logging levels in pkg/exec through the use of --vmodule=...

Fixes #164.